+++ treemmm/mroi/__init__.py
@@ -2,18 +2,22 @@
 
 from treemmm.mroi.simulator import (
     MROIResult,
+    ReallocationCurve,
     ReallocationDiagnostics,
     ReallocationPlan,
     VariableResponseCurve,
     reallocate,
+    reallocate_curve,
     simulate_mroi,
 )
 
 __all__ = [
     "MROIResult",
+    "ReallocationCurve",
     "ReallocationDiagnostics",
     "ReallocationPlan",
     "VariableResponseCurve",
     "reallocate",
+    "reallocate_curve",
     "simulate_mroi",
 ]